[autoblind] Rename binding from shadeauto to autoblind
Avoid confusion with the ShadeAuto iOS app and prevent any
implication of manufacturer involvement. The binding identity
is now "autoblind" throughout — packages, classes, thing types,
and configuration. Hardware descriptions still reference the
Norman ShadeAuto product name for discoverability.
