+++ bom/openhab-addons/pom.xml
@@ -1813,7 +1813,7 @@
     </dependency>
     <dependency>
       <groupId>org.openhab.addons.bundles</groupId>
-      <artifactId>org.openhab.binding.shadeauto</artifactId>
+      <artifactId>org.openhab.binding.autoblind</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
