+++ bundles/org.openhab.binding.autoblind/NOTICE
