+++ bundles/org.openhab.binding.autoblind/src/main/java/org/openhab/binding/autoblind/internal/api/dto/NotificationEvent.java
@@ -1,0 +1,26 @@
+/*
+ * Copyright (c) 2010-2026 Contributors to the openHAB project
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+package org.openhab.binding.autoblind.internal.api.dto;
+
+import java.util.List;
+
+/**
+ * Parsed event from the /NM/v1/notification long-poll endpoint.
+ *
+ * @author Stephen Berg (@BiloxiGeek) - Initial contribution
+ */
+public class NotificationEvent {
+
+    public long timestamp;
+    public List<Integer> peripheralList = List.of();
+}
