+++ bundles/org.openhab.binding.autoblind/src/main/java/org/openhab/binding/autoblind/internal/api/dto/RegistrationResponse.java
@@ -1,0 +1,49 @@
+/*
+ * Copyright (c) 2010-2026 Contributors to the openHAB project
+ *
+ * See the NOTICE file(s) distributed with this work for additional
+ * information.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+package org.openhab.binding.autoblind.internal.api.dto;
+
+import java.util.List;
+
+import com.google.gson.annotations.SerializedName;
+
+/**
+ * Response from the /NM/v1/registration endpoint.
+ *
+ * @author Stephen Berg (@BiloxiGeek) - Initial contribution
+ */
+public class RegistrationResponse {
+
+    @SerializedName("ThingName")
+    public String thingName = "";
+
+    @SerializedName("FirmwareVersion")
+    public String firmwareVersion = "";
+
+    @SerializedName("FirmwareId")
+    public int firmwareId;
+
+    @SerializedName("Model")
+    public String model = "";
+
+    @SerializedName("WiFiSSID")
+    public String wifiSsid = "";
+
+    @SerializedName("Peripherals")
+    public List<PeripheralStatus> peripherals = List.of();
+
+    @SerializedName("Error")
+    public int error;
+
+    @SerializedName("Timestamp")
+    public long timestamp;
+}
