+++ bundles/org.openhab.binding.autoblind/src/main/resources/OH-INF/i18n/autoblind.properties
@@ -1,0 +1,25 @@
+# add-on
+addon.autoblind.name = AutoBlind Binding
+addon.autoblind.description = Controls Norman ShadeAuto motorized shades via their local hub API.
+
+# thing types
+thing-type.autoblind.hub.label = AutoBlind Hub
+thing-type.autoblind.hub.description = Norman ShadeAuto hub connected via local network
+thing-type.autoblind.shade.label = AutoBlind Shade
+thing-type.autoblind.shade.description = A motorized shade controlled by the ShadeAuto hub
+
+# thing type config
+thing-type.config.autoblind.hub.host.label = Host
+thing-type.config.autoblind.hub.host.description = IP address or hostname of the ShadeAuto hub
+thing-type.config.autoblind.hub.pollingInterval.label = Polling Interval
+thing-type.config.autoblind.hub.pollingInterval.description = Interval in seconds between status polls
+thing-type.config.autoblind.shade.peripheralUid.label = Peripheral UID
+thing-type.config.autoblind.shade.peripheralUid.description = Unique identifier of the shade on the hub
+
+# channel types
+channel-type.autoblind.position.label = Position
+channel-type.autoblind.position.description = Shade position (0% = fully open, 100% = fully closed)
+channel-type.autoblind.batteryLevel.label = Battery Level
+channel-type.autoblind.batteryLevel.description = Battery level of the shade
+channel-type.autoblind.forceRefresh.label = Force Refresh
+channel-type.autoblind.forceRefresh.description = Clear command suppression and poll hub for current shade positions
