--- bundles/org.openhab.binding.shadeauto/src/main/java/org/openhab/binding/shadeauto/internal/ShadeAutoHandlerFactory.java
@@ -1,71 +1,0 @@
-/*
- * Copyright (c) 2010-2026 Contributors to the openHAB project
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
- */
-package org.openhab.binding.shadeauto.internal;
-
-import java.util.Hashtable;
-
-import org.eclipse.jdt.annotation.NonNullByDefault;
-import org.eclipse.jdt.annotation.Nullable;
-import org.eclipse.jetty.client.HttpClient;
-import org.openhab.binding.shadeauto.internal.discovery.ShadeAutoDiscoveryService;
-import org.openhab.binding.shadeauto.internal.handler.ShadeAutoHubHandler;
-import org.openhab.binding.shadeauto.internal.handler.ShadeAutoShadeHandler;
-import org.openhab.core.config.discovery.DiscoveryService;
-import org.openhab.core.io.net.http.HttpClientFactory;
-import org.openhab.core.thing.Bridge;
-import org.openhab.core.thing.Thing;
-import org.openhab.core.thing.ThingTypeUID;
-import org.openhab.core.thing.binding.BaseThingHandlerFactory;
-import org.openhab.core.thing.binding.ThingHandler;
-import org.openhab.core.thing.binding.ThingHandlerFactory;
-import org.osgi.service.component.ComponentContext;
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * Creates thing handlers for ShadeAuto hub and shade things.
- *
- * @author Stephen Berg (@BiloxiGeek) - Initial contribution
- */
-@NonNullByDefault
-@Component(service = ThingHandlerFactory.class, configurationPid = "binding.shadeauto")
-public class ShadeAutoHandlerFactory extends BaseThingHandlerFactory {
-
-    private final HttpClient httpClient;
-
-    @Activate
-    public ShadeAutoHandlerFactory(@Reference HttpClientFactory httpClientFactory, ComponentContext componentContext) {
-        super.activate(componentContext);
-        this.httpClient = httpClientFactory.getCommonHttpClient();
-    }
-
-    @Override
-    public boolean supportsThingType(ThingTypeUID thingTypeUID) {
-        return ShadeAutoBindingConstants.SUPPORTED_THING_TYPES_UIDS.contains(thingTypeUID);
-    }
-
-    @Override
-    protected @Nullable ThingHandler createHandler(Thing thing) {
-        ThingTypeUID thingTypeUID = thing.getThingTypeUID();
-        if (ShadeAutoBindingConstants.THING_TYPE_HUB.equals(thingTypeUID)) {
-            ShadeAutoHubHandler handler = new ShadeAutoHubHandler((Bridge) thing, httpClient);
-            bundleContext.registerService(DiscoveryService.class.getName(), new ShadeAutoDiscoveryService(handler),
-                    new Hashtable<>());
-            return handler;
-        } else if (ShadeAutoBindingConstants.THING_TYPE_SHADE.equals(thingTypeUID)) {
-            return new ShadeAutoShadeHandler(thing);
-        }
-        return null;
-    }
-}
