--- bundles/org.openhab.binding.shadeauto/src/main/java/org/openhab/binding/shadeauto/internal/api/dto/StatusResponse.java
@@ -1,34 +1,0 @@
-/*
- * Copyright (c) 2010-2026 Contributors to the openHAB project
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
- */
-package org.openhab.binding.shadeauto.internal.api.dto;
-
-import java.util.List;
-
-import com.google.gson.annotations.SerializedName;
-
-/**
- * Response from the /NM/v1/status endpoint.
- *
- * @author Stephen Berg (@BiloxiGeek) - Initial contribution
- */
-public class StatusResponse {
-
-    @SerializedName("Peripherals")
-    public List<PeripheralStatus> peripherals = List.of();
-
-    @SerializedName("Error")
-    public int error;
-
-    @SerializedName("Timestamp")
-    public long timestamp;
-}
