--- bundles/org.openhab.binding.shadeauto/src/main/resources/OH-INF/addon/addon.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<addon:addon id="shadeauto" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xmlns:addon="https://openhab.org/schemas/addon/v1.0.0"
-	xsi:schemaLocation="https://openhab.org/schemas/addon/v1.0.0 https://openhab.org/schemas/addon-1.0.0.xsd">
-
-	<type>binding</type>
-	<name>Norman ShadeAuto Binding</name>
-	<description>Controls Norman ShadeAuto motorized shades via their local hub API.</description>
-	<connection>local</connection>
-
-</addon:addon>
