--- bundles/org.openhab.binding.shadeauto/src/main/resources/OH-INF/i18n/shadeauto.properties
@@ -1,25 +1,0 @@
-# add-on
-addon.shadeauto.name = Norman ShadeAuto Binding
-addon.shadeauto.description = Controls Norman ShadeAuto motorized shades via their local hub API.
-
-# thing types
-thing-type.shadeauto.hub.label = ShadeAuto Hub
-thing-type.shadeauto.hub.description = Norman ShadeAuto Hub connected via local network
-thing-type.shadeauto.shade.label = ShadeAuto Shade
-thing-type.shadeauto.shade.description = A motorized shade controlled by the ShadeAuto Hub
-
-# thing type config
-thing-type.config.shadeauto.hub.host.label = Host
-thing-type.config.shadeauto.hub.host.description = IP address or hostname of the ShadeAuto Hub
-thing-type.config.shadeauto.hub.pollingInterval.label = Polling Interval
-thing-type.config.shadeauto.hub.pollingInterval.description = Interval in seconds between status polls
-thing-type.config.shadeauto.shade.peripheralUid.label = Peripheral UID
-thing-type.config.shadeauto.shade.peripheralUid.description = Unique identifier of the shade on the hub
-
-# channel types
-channel-type.shadeauto.position.label = Position
-channel-type.shadeauto.position.description = Shade position (0% = fully open, 100% = fully closed)
-channel-type.shadeauto.batteryLevel.label = Battery Level
-channel-type.shadeauto.batteryLevel.description = Battery level of the shade
-channel-type.shadeauto.forceRefresh.label = Force Refresh
-channel-type.shadeauto.forceRefresh.description = Clear command suppression and poll hub for current shade positions
